final-project/api/requirements.txt
amqp==5.0.2
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
attrs==20.2.0
axblack==20201012
beautifulsoup4==4.9.3
billiard==3.6.3.0
CacheControl==0.12.6
cachetools==4.1.1
celery==5.0.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
conditional==1.3
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3
cryptography==3.2.1
dataclasses==0.7
Django==3.1.1
django-crispy-forms==1.9.2
django-filter==2.4.0
django-fsm==2.7.1
django-model-utils==4.0.0
django-nose==1.4.7
djangorestframework==3.11.1
drf-firebase-auth==0.0.8
drf-yasg==1.20.0
fcm-django==0.3.7
firebase-admin==4.4.0
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-cloud-core==1.4.3
google-cloud-firestore==2.0.1
google-cloud-storage==1.32.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
graphviz==0.14.2
grpcio==1.33.2
httplib2==0.18.1
idna==2.10
ImageHash==4.1.0
importlib-metadata==2.0.0
inflection==0.5.1
interruptingcow==0.8
isort==5.6.4
itypes==1.2.0
Jinja2==2.11.2
kombu==5.0.2
lazy-object-proxy==1.4.3
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
model-bakery==1.2.1
msgpack==1.0.0
mypy-extensions==0.4.3
nose==1.3.7
numpy==1.19.4
packaging==20.4
pathspec==0.8.0
Pillow==8.0.0
prompt-toolkit==3.0.8
proto-plus==1.11.0
protobuf==3.13.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyfcm==1.4.9
pylint==2.6.0
pyparsing==2.4.7
pytz==2020.1
PyWavelets==1.1.1
redis==3.5.3
regex==2020.10.23
requests==2.24.0
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scipy==1.5.4
six==1.15.0
sortedcontainers==2.3.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
tqdm==4.50.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
zipp==3.4.0
